Deprecate `PaginatorAware` in favor of `AbstractPaginatorAware`
