+++ src/Definition/PaginatorAware.php
@@ -2,35 +2,15 @@
 
 namespace Knp\Bundle\PaginatorBundle\Definition;
 
-use Knp\Component\Pager\Paginator;
+@trigger_error('The '.__NAMESPACE__.'\PaginatorAware class is deprecated since knplabs/knp-paginator-bundle 5.x. Use '.AbstractPaginatorAware::class.' instead.', E_USER_DEPRECATED);
 
 /**
  * Class PaginatorAware.
  *
  * This is a base class that can be extended if you're too lazy to implement PaginatorAwareInterface yourself.
+ *
+ * @deprecated since knplabs/knp-paginator-bundle 5.x
  */
-abstract class PaginatorAware implements PaginatorAwareInterface
+abstract class PaginatorAware extends AbstractPaginatorAware
 {
-    /**
-     * @var Paginator
-     */
-    private $paginator;
-
-    /**
-     * Sets the KnpPaginator instance.
-     */
-    public function setPaginator(Paginator $paginator): PaginatorAwareInterface
-    {
-        $this->paginator = $paginator;
-
-        return $this;
-    }
-
-    /**
-     * Returns the KnpPaginator instance.
-     */
-    public function getPaginator(): Paginator
-    {
-        return $this->paginator;
-    }
 }
